renamed commands file
